requirements_linux_minimal.txt
# =========================
#  RecicleAI – Linux (CUDA 12.1)
# =========================

numpy==2.1.2
pandas==2.2.2
scipy==1.15.3
pillow==11.0.0
PyYAML==6.0.1
requests==2.32.3
tqdm==4.67.1
python-dateutil==2.9.0.post0
pytz==2024.1

# VISUALIZAÇÃO
matplotlib==3.10.0
seaborn==0.13.2
contourpy==1.3.1
kiwisolver==1.4.5
cycler==0.12.1
fonttools==4.52.4
packaging==24.0
pyparsing==3.1.1

# OPENCV
opencv-python==4.10.0.84

# GIT & UTILS
GitPython==3.1.45
gitdb==4.0.11
filelock==3.15.4
psutil==5.9.8
idna==3.7
urllib3==2.2.1
charset-normalizer==3.3.2
certifi==2024.2.2

# YOLOv5 DEPENDENCIES
sympy==1.13.1
mpmath==1.3.0
thop==0.1.1.post2209072238
six==1.17.0
smmap==5.0.0

# OPTIONAL
Jinja2==3.1.3
MarkupSafe==2.1.5
polars==0.20.21

# TORCH COMPATÍVEL LINUX (CUDA 12.1)
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
